Refactor balance handling to use Double instead of BigDecimal across Pocket and Transaction entities

+++ src/main/java/com/fyrdev/monyia/adapters/driven/jpa/adapter/PocketAdapter.java
@@ -43,7 +43,7 @@ public class PocketAdapter implements IPocketPersistencePort {
     }
 
     @Override
-    public int updateBalanceById(Long balance, Long pocketId) {
+    public int updateBalanceById(Double balance, Long pocketId) {
         return pocketRepository.updateBalanceById(balance, pocketId);
     }
 }
